eech-campaign: Windows x64 (MinGW) build, CI, lints
- The module builds for x86_64-pc-windows-gnu (LLP64, the DCS platform)
  and passes the whole suite under Wine, the corpus included. highlevl.h's
  WIN32 release ai_log only preprocesses under MSVC: GCC-family Windows
  builds take EECH's own non-WIN32 definition (-UWIN32, finding T1).
- The NULL-dereference outcome of a dedicated replay process is reported on
  Windows through a vectored exception handler.
- cc no longer passes -w (it silenced the -Werror= promotions); our own
  sources build with -Wall -Werror, which found %d with ptrdiff_t indices.
- CI: x86-64, i686, i686 with the original va_list idiom, -O2, Windows
  cross under Wine; rustfmt and clippy.

+++ eech-campaign/crates/eech-sys/csrc/eech_env.c
@@ -281,7 +281,7 @@ void eech_host_mobile_position (entity *en, vec3d *position)
 
 	if (!eech_current_host || !eech_current_host->mobile_position)
 	{
-		eech_abort (EECH_STATUS_UNPORTED, "mobile position", "the host provides no mobile positions (entity %d)", get_local_entity_index (en));
+		eech_abort (EECH_STATUS_UNPORTED, "mobile position", "the host provides no mobile positions (entity %d)", (int) get_local_entity_index (en));
 	}
 
 	callback_ref = eech_ref_of (en);
@@ -622,7 +622,7 @@ int assign_primary_task_to_group (entity *group_en, entity *task_en)
 	eech_last_refs[1] = eech_ref_of (task_en);
 	eech_last_ref_count = 2;
 
-	eech_abort (EECH_STATUS_BOUNDARY, boundary_message, "assign_primary_task_to_group (group %d, task %d)", get_local_entity_index (group_en), get_local_entity_index (task_en));
+	eech_abort (EECH_STATUS_BOUNDARY, boundary_message, "assign_primary_task_to_group (group %d, task %d)", (int) get_local_entity_index (group_en), (int) get_local_entity_index (task_en));
 
 	return FALSE;
 }
